Move the merging of entity updates into a utility function
